flask/requirements.txt

Flask
requests
google-generativeai
flask-cors
